2542: Remove dependencies from ToolBoxConnector to wx. Focus / refresh handling for drag and drop.
